requirements.txt
django==1.11.10
psycopg2-binary
django-govuk-template==0.4
django-govuk-forms==0.4
unittest-xml-reporting
requests
whitenoise
django-timeline-logger==1.0
django-google-analytics-app==4.2.0
django-settings-export==1.2.1
django-debug-toolbar
selenium==3.11.0
faker==0.8.13
inflect==0.2.5
